# Link our binaries with `mold` to make it faster