--- a/flake.nix
+++ b/flake.nix
@@ -132,6 +132,7 @@
             perl
             fenix-channel.rustc
             fenix-channel.clippy
+            mold
           ] ++ lib.optionals stdenv.isDarwin [
             libiconv
             darwin.apple_sdk.frameworks.Security