@justinmoon 2022-09-07
@dpc I'm getting this error after mold was merged:

```
warning: Git tree '/Users/justin/fedimint/fedimint' is dirty
error: Package ‘mold-1.2.1’ in /nix/store/0cchbcs02ggxbsy3xbmz3dl9vzv5l67w-source/pkgs/development/tools/mold/default.nix:40 is marked as broken, refusing to evaluate.

       a) To temporarily allow broken packages, you can use an environment variable
          for a single invocation of the nix tools.

            $ export NIXPKGS_ALLOW_BROKEN=1

        Note: For `nix shell`, `nix build`, `nix develop` or any other Nix 2.4+
        (Flake) command, `--impure` must be passed in order to read this
        environment variable.

       b) For `nixos-rebuild` you can set
         { nixpkgs.config.allowBroken = true; }
       in configuration.nix to override this.

       c) For `nix-env`, `nix-build`, `nix-shell` or any other Nix command you can add
         { allowBroken = true; }
       to ~/.config/nixpkgs/config.nix.
(use '--show-trace' to show detailed location information)
```

On my dreaded m1 mac