@dpc 2022-09-08
https://github.com/NixOS/nixpkgs/blob/cabc115bc2180e3c2513c4ff345e3d58305ff491/pkgs/development/tools/mold/default.nix#L13

Can you try if `nix flake update` fixes it, and push a PR if it does?